refactored the pathways folder
