--- pathways/jsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["./src/*"]
-    }
-  }
-}
-
-
-
-
-
-
-
-
-
-
