--- pathways/src/app/components/reusable/imageslider.jsx
@@ -1,52 +1,0 @@
-"use client";
-import { useState } from "react";
-import Image from "next/image";
-import photo from "../../assets/image/fdiphoto.png";
-// import photo from "@/assets/image/fdiphoto.png";
-
-const images = [
-  { src: photo, alt: "first slide" },
-  { src: photo, alt: "second slide" },
-  { src: photo, alt: "third slide" },
-];
-
-export default function ImageSlider() {
-  const [currentIndex, setCurrentIndex] = useState(0);
-
-  const prevSlide = () => {
-    setCurrentIndex((prev) => (prev === 0 ? images.length - 1 : prev - 1));
-  };
-
-  const nextSlide = () => {
-    setCurrentIndex((prev) => (prev === images.length - 1 ? 0 : prev + 1));
-  };
-
-  return (
-    <div className="relative w-screen h-screen overflow-hidden mt-20">
-      <div
-        className="flex transition-transform duration-500 h-full"
-        style={{ transform: `translateX(-${currentIndex * 100}%)` }}
-      >
-        {images.map((img, idx) => (
-          <div key={idx} className="relative w-screen h-screen flex-shrink-0">
-            <Image src={img.src} alt={img.alt} fill className="object-cover" />
-          </div>
-        ))}
-      </div>
-
-      {/* Navigation buttons */}
-      <button
-        onClick={prevSlide}
-        className="absolute left-2 top-1/2 -translate-y-1/2 bg-black/40 text-white p-2 rounded-full"
-      >
-        &#8592;
-      </button>
-      <button
-        onClick={nextSlide}
-        className="absolute right-2 top-1/2 -translate-y-1/2 bg-black/40 text-white p-2 rounded-full"
-      >
-        &#8594;
-      </button>
-    </div>
-  );
-}
