--- pathways/src/app/layout.jsx
@@ -1,35 +1,0 @@
-import { DM_Sans } from "next/font/google";
-import "./globals.css";
-import Header from "./components/reusable/header";
-import Footer from "./components/reusable/footer";
-
-const dmSans = DM_Sans({
-  variable: "--font-dm-Sans",
-  subsets: ["latin"],
-});
-
-// const geistMono = Geist_Mono({
-//   variable: "--font-geist-mono",
-//   subsets: ["latin"],
-// });
-
-export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="en">
-      <body className={`${dmSans.variable} font-['dm-sans'] antialiased`}>
-        <div className="relative w-screen h-screen overflow-x-hidden max-w-[90rem] flex flex-col">
-          <Header />
-          {children}
-          <Footer />
-        </div>
-      </body>
-    </html>
-  );
-}
-
-// items-center
