+++ src/app/contact/page.jsx
@@ -6,15 +6,15 @@ import hands from "../assets/image/groupie.png";
 
 export default function Contact() {
   return (
-    <div className="mt-[10rem] border-t border-gray-300">
-      <div className=" text-sm text-gray-500 flex items-center space-x-2 mx-18 mt-10 mb-6">
+    <div className="md:mt-[10rem] mt-[5rem] border-t border-gray-300">
+      <div className=" text-sm text-gray-500 flex items-center space-x-2 md:mx-18 mx-6 mt-10 mb-6">
         <Link href="/">Home</Link>
         <span>→</span>
-        <span className="text-black font-medium">Team</span>
+        <span className="text-black font-medium">Contact Us</span>
       </div>
-      <div className="max-w-6xl mx-auto grid grid-cols-1 md:grid-cols-2 gap-12 ">
+      <div className="max-w-6xl md:mx-auto mx-4 grid grid-cols-1 md:grid-cols-2 gap-12 ">
         <div>
-          <p className="text-2xl md:text-4xl font-bold leading-relaxed mb-6">
+          <p className="text-3xl md:text-4xl font-bold leading-relaxed mb-6">
             Have questions or <br />
             interested in learning more <br />
             about FPDI?
